sock.on functions transferred to events/socketEvents
