+++ .eslintrc.json
@@ -9,7 +9,5 @@
         "ecmaVersion": 2020
     },
     "rules": {
-        "semi": ["error", "always"],
-        "quotes": ["error", "single"]
     }
 }
